Simplify tests with more fixtures
